Modify: Carousel 추가 laptop-width
