+++ src/components/Pledge/Candidate/CandidateSection.tsx
